+++ src/lib/prisma.ts
@@ -7,14 +7,12 @@ const globalForPrisma = globalThis as unknown as {
 export const prisma =
     globalForPrisma.prisma ??
     new PrismaClient({
-        datasourceUrl: process.env.DATABASE_URL,
         log:
             process.env.NODE_ENV === "development"
-                ? ["query", "error", "warn"]
+                ? ["error", "warn"]
                 : ["error"],
     });
 
 if (process.env.NODE_ENV !== "production") globalForPrisma.prisma = prisma;
 
 export default prisma;
-